feat: implement dominator and post-dominator analyses
This introduces the incrementally updateable dominator and
post-dominator analyses from LLVM, based on the Semi-NCA algorithm. This
enables us to keep the (post-)dominator tree analysis up to date during
rewriting of the IR, without needing to recompute it from scratch each
time.

This is an initial port of the original C++ code, modified to be more
Rust-like, while still remaining quite close to the original. This could
be easily cleaned up/rewritten in the future to be more idiomatic Rust,
but for now it should suffice.