feat: implement loop analysis
This commit implements the generic loop analysis from LLVM, in Rust. It
is a more sophisticated analysis than the one in HIR1, and provides us
with some additional useful information that will come in handy during
certain transformations to MASM.

See the "Loop Terminlogy" document on llvm.org for details on how LLVM
reasons about loops, which corresponds to the analysis above.