feat: implement liveness analysis, improvements to dataflow/analysis
This commit ports over the liveness analysis from HIR1, and makes
various improvements/extensions to the dataflow framework. The primary
changes to the dataflow framework are the following:

* Give the solver a reference to an AnalysisManager so that analyses can
  be requested by a DataFlowAnalysis implementation during
  initialization. We primarily are interested in this to allow obtaining
  cached analyses from it, as the alternative is computing those
  analyses from scratch when needed, e.g. dominator tree, loop fores.
* Use the dominator tree and loop forest analyses, when applicable, to
  choose a better visitation order when initializing dense dataflow
  analyses. For forward analysis, we follow the CFG in reverse postorder,
  and skip any unreachable blocks entirely. For backward analysis, we
  visit the CFG in postorder, also skipping unreachable blocks. If an op
  has no regions, or the regions do not use SSA dominance, or consist
  only of a single block, the naive visitation order is used, i.e. just
  iterating over the set of regions/blocks/ops regardless of CFG edges
  between them.
* An additional hook was added to the dense forward/backward dataflow
  analysis traits that allows an analysis to control how propagation is
  handled along unstructured/non-regional control flow edges. This was
  needed by liveness analysis to be able to handle subtleties in the
  transfer function along such edges.

Liveness analysis has been adapted to be run as a dense-backward
dataflow analysis via the dataflow framework, and to support
region-based (structured) control flow in addition to unstructured
control flow. With some additional tweaks, support for inter-procedural
analysis could be supported as well, which would primarily be useful to
for optimizing calls to functions where the caller is the sole caller of
the function, and not all of the arguments are being used (and thus
could be removed from both the callsite and the callee definition).
Combined with sparse-conditional constant propagation, it now also takes
into account whether or not control flow edges are executable.