feat: initial draft of control flow lifting transform
This commit implements a generic control flow lifting transformation
framework, and an implementation of it for the HIR dialect.

The tranform converts unstructured control flow operations to structured
control flow ops, e.g. `hir.cond_br` to `hir.if`. This includes complex
control flow such as loops with loop-carried variables, reduce-style
loops, etc.

Not all control flow is guaranteed to be lifted however. Currently, the
primary limitation is that all return-like operations in a region must
be equivalent (i.e. the same operation type), otherwise some control
flow will be left unstructured.

For now, on the backend, this will result in a legalization error, as we
do not have lowerings for unstructured control flow ops at this time. In
the future we plan to lower directly to structured control flow, so
this is more of a temporary solution for us anyway; but may also be
useful for other compiler frontends.