refactor(ir): rework handling of entities with parents
This commit refactors the way entities with parents store and maintain
their parent/child relationships. In particular, such entities must now
implement a trait, which identifies the type of parent for a given
entity. Parents must also implement a twin trait that provides us with
the offset from the parent type to the EntityList in which a specific
child type is to be tracked.

This permits two key benefits:

1. We do not need to borrow an entity to traverse to its parent, this
   vastly reduces the amount of borrowing required to traverse and modify
   the IR, which significantly cuts down the chances of borrowing
   conflicts at runtime.
2. We can automatically link/unlink parents when an EntityList is
   modified (i.e. on insertion the parent link is set, and on removal it
   is unlinked).

The second benefit is particularly noticeable, because we no longer need
to carefully ensure that we maintain those relationships manually every
time we modify an entity, remove it from its parent, transfer to a new
parent, etc.

There are still two cases where further refactoring is needed to bring
these benefits to _all_ entity types that are stored in an EntityList:

1. Some entities may have multiple parent types from a limited set. For
   example, OpOperand can have as parent either BlockArgument or
   OpResult, and the current EntityWithParent and EntityParent traits do
   not provide for entity relationships of this kind yet. For now, we
   get close enough with just the EntityListItem callbacks and the
   StorableEntity trait impl to be able to replicate the behavior we'd
   otherwise get, but it would be preferable to be able to unify this if
   we can.
2. Some entities have a concrete Op type as their parent, and can belong
   to any number of them. SymbolUse is an example of this, as only
   concrete Op implementations know whether or not the operation
   references symbols, and thus requires a SymbolUseList. This is a
   harder case to solve for, as we cannot rely on the type system here
   the way we currently are. Because we would not be able to look up
   a parent list given a child type, it would be necessary for entity
   lists to be aware of the address of the entity in which they are
   stored. This would inflate the size of the EntityList type, and be
   tricky to orchestrate, as we would have to ensure that whenever an
   entity is created, that once it has a stable address we notify all of
   its member entity lists. For now, SymbolUse (and any other
   as-yet-undefined entities like it) must be manually managed as
   before.

Amongst other changes, this commit attempts (and I think succeeds, so
far as I can tell) at removing any unnecessary borrows related to
traversing parents/grandparents, as well as simplifying various bits of
code related to managing parent/child relationships.