test(hir): implement complex control flow test with early exits
This test, beyond just evaluating the control flow lifting
transformation on a complex control flow example, also evaluates the
application of the Canonicalization pass, and a subset of the
canonicalization patterns on our structured control flow ops.

The control flow example here consists of nested loops with
conditionals, with an early return in the nested loop. Lifting this into
structured control flow requires exercising the most complicated parts
of the CFG-to-SCF transformation pass, as it must ensure that the early
return is handled in a structured fashion.