refactor(dataflow): support get_or_create of analysis state without mutable borrowing
This is primarily necessary in sparse dataflow analysis, whose anchors
are predominantely values, and when working with op operands, the same
analysis state may be requested multiple times, but only for read-only
uses. The previous way we fetched the analysis state required a brief
mutable borrow of the state, but if the same analysis state was already
requested for a previous operand, this would cause a runtime panic when
the alias is detected.

The only reason we required the mutable borrow just to get_or_create
analysis state, was because in order to subscribe an analysis to that
state, we required a mutable reference to the underlying
AnalysisStateInfo. By protecting that with a RefCell, we can sidestep
this, and we're never going to have a case where we're borrowing the
same AnalysisStateInfo subscriptions at the same time. This allows us to
then avoid the mutable borrow of the analysis state entirely, and make a
cleaner distinction between read/write and readonly AnalysisState
guards.