fix(ir): dominance info initialization incomplete
This commit fixes an issue that was causing dominance information to be
incorrect due to incomplete metadata attached to dominance tree nodes,
specifically:

1. The in/out visitation order
2. The dominance tree level/depth

These pieces of information are supposed to be computed during
construction and upon request, but the implementation of 2 was
incomplete, and the function that performs 1 was never called. Both
issues have now been fixed.

Additionally, an explicit request to update the DFS visitation order
during the CFG-to-SCF transformation was added, due to the
transformation making many requests for that information, and because
the dominance info is being regularly invalidated during the transform.
This new call ensures that it is updated before it is needed, and
remains valid within that scope.

Both of these issues address the outstanding compiler bug that was
causing the is_prime test to fail during execution by improperly
returning poison. The specific issue is that, during the control flow
lifting transformation, and specifically within the part that transforms
cycles in the CFG into a reduce-form loop, a check is performed to see
if any uses of certain values will escape the loop body, which is,
crucially, based on dominance information. If a value can escape, a new
block parameter must be introduced in the loop header, latch block, and
the block containing each escaping use. If it is determined that a
predecessor edge to the latch block is not dominated by the loop header,
then the value passed to the newly introduced block parameter at the
site of the use, is set to undef/poison, as by definition, there can be
no valid uses of the original value along that edge. However, because
our dominance information was incorrect, it was incorrectly flagging
certain edges as not being dominated, when they in fact were, and thus
silently rewriting valid uses with undef/poison that were then observed
at runtime.