ci: check each workspace member for build errors
Due to the workspace feature unification `cargo make check`
may not catch build errors that occur during the publishing workflow as
a result of the release process running `cargo build` on each workspace
member as a standalone crate, thus resulting in a different set of
default features. This was causing releases to fail for the
`midenc-session` crate.

This commit adds a new `cargo make check-each` task which runs `cargo
check` on each workspace member independently, rather than on the
workspace level as `cargo make check` is configured to do. This new
task is then used in the CI workflow to ensure we don't miss any
required features of our dependencies.