fix(cfg-to-scf): incorrect folding of continuation region blocks
This bug was reported in #510, and caused an infinite loop in the
CFG-to-SCF transformation pass, due to misclassifying the structure of
the IR when matching Case 3 as described in the inline documentation of
the `transform_to_structured_cf_branches` function. The resulting
misclassification was due to mistakenly identifying a multi-exit
continuation region as a single-exit continuation, thereby resulting in
a cycle in the transformation.

Fixes #510