fix: delayed registration of scf dialect causes canonicalizations to be skipped
When the `Canonicalizer` pass is initialized by the `PassManager`, it
visits each of the dialects registered with the `Context` at that time,
and populates the set of canonicalization patterns for all operations
registered with that dialect. Once constructed, any dialect
registrations after that point will not have the canonicalization
patterns of their operations included in the set applied by the
`Canonicalizer`.

The registration of the SCF dialect was occuring implicitly, when the
`CFGToSCFTransform` pass executed and encountered a `cf` dialect
operation to transform. This is well after the point where the
compiler's pass pipeline is instantiated and initialized. As a result,
the canonicalization patterns of the SCF dialect were never registered
with the `Canonicalizer`, because the SCF dialect itself was not
registered at the point where the `Canonicalizer` performs its own
initialization.

This change modifies the two existing transform passes such that they
force registration of the dialects they operate on. Really only the
`CFGToSCFTransform` pass is relevant here, as it is the only current
dialect that has its registration delayed so long, which is a short-term
consequence of the frontend not emitting any SCF ops.

Closes #563