fix(scf): over-eager trivial if-to-select rewrite
This rewrite pattern was not considering conversions of scf.if
operations for which all yielded results were hoisted, but the body of
one or more of the scf.if regions still contained operations with side
effects that were not affected by the rewrite. The transform would fail
as a result.