fix: start `Linker::reserved_memory_pages` on the next page after rustc
reservation for the Rust stack.
This commit fixes the bug that manifests as if a `Vec` was allocated at
the same address as the Rust stack. The root cause was that we put the
first global variable (`__stack_pointer`) at the same address as the Rust
stack pointer. Since the value of the `__stack_pointer` is the Rust stack
address and the `BumpAlloc` instance is allocated on the stack, the `BumpAlloc::top`
pointer got the value of the Rust stack address (`__stack_pointer`) and
`BumpAlloc::maybe_init()` constructor skipped the `BumpAlloc::top` initialization.