fix: convert `auth-*` function names to snake case (`auth_*`)
The miden-base expects the authentication component expects an authentication procedure with
the name `auth_*` (underscore). Since WIT names are in kebab case convert this one to snake
case. In the future miden-base will use annotation to mark the procedure as authentication
procedure and we will remove this. Until then we use the following workaround.