build: switch workspace crates to crates.io version of midenc-hir-type
Now that midenc-hir-type is used by miden-assembly-syntax, and its types
are re-exported there - it is necessary for compiler crates to depend on
the same version as miden-assembly-syntax to allow for passing those
types between APIs of the compiler, miden-assembly-syntax, and
miden-mast-package. If we instead retained a direct path dependency on
the source of midenc-hir-type, Cargo would see the dependencies as
effectively different versions, and therefore make it impossible to pass
types between those crates as they would not be considered the same
anymore. Manual conversion would be possible, but undesirable.

Since we plan to split out midenc-hir-type eventually anyway (or at the
very least, do what this commit does and treat it as an entirely
separate crate from the rest of the compiler), this change makes that
transition easier, at the expense of being a bit confusing when making
changes to midenc-hir-type that need to be propagated to crates.io first
in order to meaningfully test the full compiler pipeline with those
changes.