Change the memory representation of 64-bit, dual-limbed values to be little-endian.
In memory the low (least significant) 32-bit limb is now loaded or
stored from/to `addr` and the high 32-bit limb from/to `addr+1`.  On the
stack the limb order is still big-endian.

For testing the `ToMidenRepr` array of felts are now little-endian.  The
`FromMidenRepr` is still a big-endian array of felts, as per the stack.