fix: rust-lld fails due to divergent definitions of __rdl_alloc_error_handler
It appears that Rust for the wasm32-* targets now emits a second object
file containing stub definitions for the default #[alloc_error_handler]
that is defined by liballoc/libstd (the implementations are different
depending on whether a crate is #![no_std] or not). The problem is that
the signature of these stubs does not match the signature of the actual
definitions, i.e. it is `fn() -> !` rather than `fn(core::alloc::Layout)
-> !`. Because these stubs are fed to rust-lld, we get conflicting
symbol errors due to the stub definition conflicting with the actual
definition in liballoc.

It isn't clear to me why this second object file is emitted by Rust, nor
why the signature is wrong. My assumption is that it has something to do
with the fact that we're producing a `cdylib` artifact, or perhaps due
to our use of `cargo -Zbuild-std` (more likely), but I wasn't able to
confirm either option.

Regardless, the fix appears to be to ensure that we define our own
`#[alloc_error_handler]` with the correct signature in the crate being
compiled. This ensures that only one definition is present, ours, and
makes the conflict go away. The downside of this (currently), is that it
also requires activation of `#[feature(alloc_error_handler)]`, which
cannot be hidden by our proc-macro machinery that hides boilerplate.

My hope is that we can figure out the actual underlying issue here, or a
fix is shipped in a new nightly and we can make this workaround go away.