fix(codegen): emit operand scheduling code in switch lowering as needed
This commit modifies the binary search lowering strategy for
switch ops so that we emit operand scheduling code when the contents of
the operand stack diverge when yielding from the then/else regions. This
aligns it with the lowering for `scf.if` which does the same thing.

Previously, we were simply asserting that the two regions do not
diverge, but that is not guaranteed to be the case, and while we
addresssed that for `scf.if`, we never did the same for
`scf.index_switch`.

Fixes #827